translator/requirements.txt
fastapi
uvicorn
pydantic
requests
transformers
sacremoses
dvc
dvc-gdrive
sentencepiece
transformers[torch]
httpx
pytest
allure-pytest